Part B finished. DCM left not working, Antibounce is in a state of demonstration
The 4 led display  units are fed information correctly. Antibounce works
correctly. To proceed to the next part Antibounce has to change
clk_count to count 2^20 cycles instead of 2^6 (i dont need to see result
on testbench anymore) , system doesnt need to show antibounce stabilized
button and button anymore.  DCM has to properly divide the time and feed
that divided time to ledDataFeeder in order for it to work slowly on the
FPGA